Revert "feat(plugins): Plugin binary store service (#714)"
This reverts commit 767a07cd62946ac313f465adaddf4180451653b8.
